// Downloads/frontend2-test/Downloads/FRONNNT-test2 (1)/FRONNNT-test2/src/app/Models/cdc.model.ts
import { Site } from "./Sites.model";
import { Dev } from "./dev.model";
import { Fonction } from "./fonction.model";
import { Reverse } from "./reverse.model";

export class CDC {
    idCdc: any;
    nomCdc: any;
    refCdc: any;
    indCdc: any;
    creation: any; 
    signatureOk: any;
    refCdcArdia: any;
    dev: any; 
    sites:any;
    reverse: any; 
    fonctions:any;
  }